Updated website files, packages application into EXE, zipped files for download
