+++ HTML/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<meta charset = "utf-8">
-  	<title>Group Project</title>
+  	<title>G3 Consulting - Home</title>
   	<link rel="tab icon" type="image/x-icon" href="IMG_3906.png">
   	<link rel="stylesheet" href="stylesheet.css">
 </head>
@@ -12,22 +12,22 @@
 		<nav>
 			<ul>
 				<li><a href="index.html">Home</a></li>
-				<li><a href="projectOne.html">Project 1</a></li>
-				<li><a href="projectTwo.html">Project 2</a></li>
-				<li><a href="projectThree.html">Project 3</a></li>
+				<li><a href="projectOne.html">Product 1</a></li>
+				<li><a href="projectTwo.html">Product 2</a></li>
+				<li><a href="projectThree.html">Product 3</a></li>
   			</ul>
 		</nav>	
 	</header>
 
-	<h1>Who is Team 3?</h1>
-	<p>Team 3 consists of Junior and Senior Wayne State University students, Ashiq Ahmed, Mina Stanton, and McKayla Kleinschrodt. These students have been paired together in Software Eningeering, CSC 4110, to create applications that teach skills that will translate to the professional world.</p>
+	<h1>Who is G3 Consulting?</h1>
+	<p>G3 Constuling consists of Junior and Senior Wayne State University students, Ashiq Ahmed, Mina Stanton, and McKayla Kleinschrodt. These students have been paired together in Software Eningeering, CSC 4110, to create applications that teach skills that will translate to the professional world.</p>
 	<a class="nav-link" href="GetData.zip">Product Download</a>
 
 
 <footer>
 	<div class="foot">
 	<small>
-  	        <p>Group Project: Team 3 <time datetime="2022 - Winter Semester">
+  	        <p>G3 Consulting<time datetime="2022 - Winter Semester">
 			</time><a href= "mailto: kaylakleinschrodt@wayne.edu">
 	</small>
 	<div>
