+++ HTML/projectOne.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<meta charset = "utf-8">
-  	<title>Group Project 1</title> 
+  	<title>G3 Consulting - Employee Database</title> 
   	<link rel="tab icon" type="image/x-icon" href="favicon.ico">
   	<link rel="stylesheet" href="stylesheet.css">
 </head>
@@ -12,14 +12,14 @@
 		<nav>
 			<ul>
 				<li><a href="index.html">Home</a></li>
-				<li><a href="projectOne.html">Project 1</a></li>
-				<li><a href="projectTwo.html">Project 2</a></li>
-				<li><a href="projectThree.html">Project 3</a></li>
+				<li><a href="projectOne.html">Product 1</a></li>
+				<li><a href="projectTwo.html">Product 2</a></li>
+				<li><a href="projectThree.html">Product 3</a></li>
   			</ul>
 		</nav>	
 	</header>
 
-	<h1>Project 1 -- Employee Database</h1>
+	<h1>Product 1 -- Employee Database</h1>
 	<p>Description: The following .exe file will run a program that will allow users to add, search, and display a database of employee records. This database uses pickling technology that allows for data to be consistent regardless of the recent status of the program.</p>
 
 	<a class="nav-link" href="/Project One/Project_One.zip">Product Download<a>
@@ -30,7 +30,7 @@
 <footer>
 	<div class="foot">
 	<small>
-  	        <p>Group Project: Team 3 <time datetime="2022 - Winter Semester">
+  	        <p>G3 Consulting<time datetime="2022 - Winter Semester">
 			</time><a href= "mailto: kaylakleinschrodt@wayne.edu">
 	</small>
 	<div>
