+++ HTML/projectTwo.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<meta charset = "utf-8">
-  	<title>Group Project</title>
+  	<title>G3 Consulting - BarMaster</title>
   	<link rel="tab icon" type="image/x-icon" href="favicon.ico">
   	<link rel="stylesheet" href="stylesheet.css">
 </head>
@@ -12,19 +12,25 @@
 		<nav>
 			<ul>
 				<li><a href="index.html">Home</a></li>
-				<li><a href="projectOne.html">About</a></li>
-				<li><a href="projectTwo.html">Experience</a></li>
-				<li><a href="projectThree.html">Studies</a></li>
+				<li><a href="projectOne.html">Product 1</a></li>
+				<li><a href="projectTwo.html">Product 2</a></li>
+				<li><a href="projectThree.html">Product 3</a></li>
   			</ul>
 		</nav>	
 	</header>
 
+	<h1>Product 2 -- BarMaster Application</h1>
+	<p>Description: The following .exe file will run a program that will allow users to place orders for multiple drinks using a tkinter GUI. The application will also allow users to manipulate inventory values when needed.</p>
+
+	<a class="nav-link" href="/Project One/Project_Two.zip">Product Download<a>
+	<p>User may not sell or distribute data without the express consent of Forestview, by downloading this file, you are waiving your rights to legal action is data is breached!
+	Security checks will periodically be performed on the above software as needed.</p>
 
 
 <footer>
 	<div class="foot">
 	<small>
-  	        <p>Group Project: Team 3 <time datetime="2022 - Winter Semester">
+  	        <p>G3 Consulting <time datetime="2022 - Winter Semester">
 			</time><a href= "mailto: kaylakleinschrodt@wayne.edu">
 	</small>
 	<div>
