--- Project Two/Inventory.json
@@ -1,1 +1,0 @@
-{"vodka": 997, "gin": 999, "tequila": 991, "rum": 993, "ginger_beer": 993, "triple_sec": 985, "cream": 996, "juice": 989, "soda_water": 995, "tea": 999, "lemonade": 999, "grenadine": 999, "fruit": 986, "ice": 996, "egg": 999}
