--- Project Two/systemLog.json
@@ -1,1 +1,0 @@
-[{"time": "10:14pm", "drink": "cosmo", "transaction": "Server 2 served a cosmo"}, {"time": "04/02/2022, 10:15:01"}, {"drink": "cosmo"}, {"transaction": "At 1648908901.530048, Employee 2 served a Cosmo"}, {"time": "04/02/2022, 10:22:46"}, {"drink": "cosmo"}, {"transaction": "At 2022-04-02 10:22:46.748125, Employee 2 served a Cosmo"}, {"time": "04/02/2022, 12:43:03"}, {"drink": "cosmo"}, {"transaction": "At 2022-04-02 12:43:03.908058, Employee 3 served a Cosmo"}, {"time": "04/05/2022, 15:35:43"}, {"drink": "cosmo"}, {"transaction": "At 2022-04-05 15:35:43.155385, Employee 0 served a Cosmo"}, {"time": "04/05/2022, 15:37:32"}, {"drink": "mocktail"}, {"transaction": "At 2022-04-05 15:37:32.835605, Employee 2 served a Island Mocktail"}, {"time": "04/05/2022, 15:42:22"}, {"drink": "marg"}, {"transaction": "At 2022-04-05 15:42:22.721740, Employee 0 served a Margarita"}, {"time": "04/05/2022, 15:42:30"}, {"drink": "pina"}, {"transaction": "At 2022-04-05 15:42:30.727497, Employee 7 served a Pina Colada"}]
